requirements.txt
fastapi>=0.100.0,<0.101.0
uvicorn>=0.15.0,<0.16.0
pydantic>=2.0.0,<3.0.0
pydantic-settings>=2.0.0
mysql-connector-python>=8.0.26,<8.1.0
python-dotenv>=0.21.0
pytest>=7.0.0
httpx>=0.24.0
pytest-asyncio>=0.21.0
gunicorn==20.1.0
